refactor(Product): 상품 패키지 분리
